Completed Value Added Features and Refactors

+++ HotelListing.API/Controllers/CountriesV2Controller.cs
@@ -6,24 +6,21 @@ using AutoMapper;
 using HotelListing.API.Contracts;
 using Microsoft.AspNetCore.Authorization;
 using HotelListing.API.Core.Exceptions;
-using HotelListing.API.Core.Models;
-using static HotelListing.API.Core.Models.QueryParameters;
-using Microsoft.OData.ModelBuilder;
-using System.Diagnostics.Metrics;
+using Microsoft.AspNetCore.OData.Query;
 
 namespace HotelListing.API.Controllers
 {
     [Route("api/v{version:apiVersion}/countries")]
     [ApiController]
-    [ApiVersion("1.0", Deprecated = true)]
-    public class CountriesController : ControllerBase
+    [ApiVersion("2.0")]
+    public class CountriesV2Controller : ControllerBase
     {
         // private readonly HotelListingDbContext _context;
         private readonly IMapper _mapper;
         private readonly ICountriesRepository _countriesRepository;
         private readonly ILogger<CountriesController> _logger;
 
-        public CountriesController(IMapper mapper, ICountriesRepository countriesRepository,
+        public CountriesV2Controller(IMapper mapper, ICountriesRepository countriesRepository,
             ILogger<CountriesController> logger)
         {
             // _context = context;
@@ -32,26 +29,15 @@ namespace HotelListing.API.Controllers
             this._logger = logger;
         }
 
-        // GET: api/Countries/GetAll
-        [HttpGet("GetAll")]
+        // GET: api/Countries
+        [HttpGet]
+        [EnableQuery]
         public async Task<ActionResult<IEnumerable<GetCountryDto>>> GetCountries()
         {
             // var countries = await _context.Countries.ToListAsync();
-
-            //var countries = await _countriesRepository.GetAllAsync();
-            //var records = _mapper.Map<List<GetCountryDto>>(countries);
-            //return Ok(records);
-
-            var countries = await _countriesRepository.GetAllAsync<GetCountryDto>();
-            return Ok(countries);
-        }
-
-        // GET: api/Countries/?StartIndex=0&pagesize=25&PageNumber=1
-        [HttpGet]
-        public async Task<ActionResult<PagedResult<GetCountryDto>>> GetPagedCountries([FromQuery] QueryParameters queryParameters)
-        {
-            var pagedCountriesResult = await _countriesRepository.GetAllAsync<GetCountryDto>(queryParameters);
-            return Ok(pagedCountriesResult);
+            var countries = await _countriesRepository.GetAllAsync();
+            var records = _mapper.Map<List<GetCountryDto>>(countries);
+            return Ok(records);
         }
 
         // GET: api/Countries/5
@@ -60,16 +46,16 @@ namespace HotelListing.API.Controllers
         {
             //var country = await _context.Countries.Include(q => q.Hotels)
             //    .FirstOrDefaultAsync(q => q.Id == id);
+            var country = await _countriesRepository.GetDetails(id);  
 
-            var country = await _countriesRepository.GetDetails(id);
-            //if (country == null)
-            //{
-            //    throw new NotFoundException(nameof(GetCountry), id);
-            //}
-            //var countryDto = _mapper.Map<CountryDto>(country);
-            //return Ok(countryDto);
+            if (country == null)
+            {
+                throw new NotFoundException(nameof(GetCountry), id);
+            }
 
-            return Ok(country);
+            var countryDto = _mapper.Map<CountryDto>(country);
+
+            return Ok(countryDto);
         }
 
         // PUT: api/Countries/5
@@ -84,22 +70,19 @@ namespace HotelListing.API.Controllers
             }
 
             // var country = await _context.Countries.FindAsync(id);
+            var country = await _countriesRepository.GetAsync(id);
 
-            //var country = await _countriesRepository.GetAsync(id);
-            //if (country == null)
-            //{
-            //    throw new NotFoundException(nameof(GetCountry), id);
-            //}
-            //_mapper.Map(updateCountryDto, country);
+            if (country == null)
+            {
+                throw new NotFoundException(nameof(GetCountry), id);
+            }
+
+            _mapper.Map(updateCountryDto, country);
 
             try
             {
                 //await _context.SaveChangesAsync();
-
-                //await _countriesRepository.UpdateAsync(country);
-
-                await _countriesRepository.UpdateAsync(id, updateCountryDto);
-
+                await _countriesRepository.UpdateAsync(country);
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -122,17 +105,13 @@ namespace HotelListing.API.Controllers
         [Authorize]
         public async Task<ActionResult<Country>> PostCountry(CreateCountryDto createCountryDto)
         {
+            var country = _mapper.Map<Country>(createCountryDto);
 
             //await _context.Countries.AddAsync(country);
-
             //await _context.SaveChangesAsync();
+            await _countriesRepository.AddAsync(country);
 
-            //var country = _mapper.Map<Country>(createCountryDto);
-            //await _countriesRepository.AddAsync(country);
-            //return CreatedAtAction("GetCountry", new { id = country.Id }, country);
-
-            var country = await _countriesRepository.AddAsync<CreateCountryDto, GetCountryDto>(createCountryDto);
-            return CreatedAtAction(nameof(GetCountry), new { id = country.Id }, country);
+            return CreatedAtAction("GetCountry", new { id = country.Id }, country);
         }
 
         // DELETE: api/Countries/5
@@ -141,17 +120,16 @@ namespace HotelListing.API.Controllers
         public async Task<IActionResult> DeleteCountry(int id)
         {
             //var country = await _context.Countries.FindAsync(id);
-
-            //var country = await _countriesRepository.GetAsync(id);
-            //if (country == null)
-            //{
-            //    throw new NotFoundException(nameof(GetCountry), id);
-            //}
+            var country = await _countriesRepository.GetAsync(id);
+            if (country == null)
+            {
+                throw new NotFoundException(nameof(GetCountry), id);
+            }
 
             //_context.Countries.Remove(country);
             //await _context.SaveChangesAsync();
-
             await _countriesRepository.DeleteAsync(id);
+
             return NoContent();
         }
